pyproject.toml
[project]
name = "bosty-radio"
version = "0.1.0"
description = "Headless GPIO Rotary Internet Radio for Raspberry Pi"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "textual>=0.68.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "RPi.GPIO>=0.7.1; sys_platform == 'linux'",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "linkchecker"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []

[tool.black]
line-length = 100
target-version = ['py312']

[tool.ruff]
line-length = 100
target-version = "py312"

